Feat/payble full (#14)
* payble oracle

* forward fees

* fix test

* add unit tests

* implement payments for oracle

* swap functionality

* typo

* finish

+++ src/example/tokens/fUSDC.sol
@@ -12,7 +12,7 @@ contract FUSDC is ERC20, Ownable {
         return 6;
     }
 
-    function mint(address account, uint value) external onlyOwner {
+    function mint(address account, uint256 value) external onlyOwner {
         _mint(account, value);
     }
 }
