+++ src/example/tokens/fWETH.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.16;
 import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract FWETH is ERC20 {
-    event Deposit(address indexed dst, uint val);
-    event Withdrawal(address indexed src, uint val);
+    event Deposit(address indexed dst, uint256 val);
+    event Withdrawal(address indexed src, uint256 val);
 
     error TransferFailed();
 
@@ -25,7 +25,7 @@ contract FWETH is ERC20 {
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint val) public {
+    function withdraw(uint256 val) public {
         (bool callSuccess, ) = msg.sender.call{value: val}("");
 
         if (!callSuccess) {
